+++ .changeset/config.json
@@ -6,5 +6,5 @@
   "access": "public",
   "baseBranch": "main",
   "updateInternalDependencies": "patch",
-  "ignore": ["demo", "docs"]
+  "ignore": ["sandbox", "docs"]
 }
