+++ apps/sandbox/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "demo",
+  "name": "sandbox",
   "type": "module",
   "version": "0.0.1",
   "scripts": {
