+++ apps/sandbox/src/components/examples/DerivedExample.tsx
