+++ apps/sandbox/src/components/examples/SharedExample.tsx
