+++ apps/sandbox/src/components/ui/checkbox.tsx
@@ -2,7 +2,7 @@ import * as CheckboxPrimitive from "@radix-ui/react-checkbox";
 import { CheckIcon } from "lucide-react";
 import * as React from "react";
 
-import { cn } from "@apps/demo/src/lib/utils";
+import { cn } from "@apps/sandbox/src/lib/utils";
 
 function Checkbox({
   className,
