+++ apps/sandbox/src/components/ui/tabs.tsx
@@ -1,7 +1,7 @@
 import * as TabsPrimitive from "@radix-ui/react-tabs";
 import * as React from "react";
 
-import { cn } from "@apps/demo/src/lib/utils";
+import { cn } from "@apps/sandbox/src/lib/utils";
 
 function Tabs({
   className,
