+++ apps/sandbox/src/layouts/Layout.astro
@@ -1,7 +1,7 @@
 ---
 import '../styles/global.css';
 import Nav from '../components/Nav.tsx';
-import { links } from '../lib/constants';
+import { links } from '../lib/constants.ts';
 
 interface Props {
   hideNav?: boolean;
