+++ apps/sandbox/src/pages/vanilla.astro
