+++ package.json
@@ -9,8 +9,7 @@
   },
   "scripts": {
     "build": "pnpm -F \"./packages/*\" build",
-    "demo:dev": "pnpm -F demo dev",
-    "demo:build": "pnpm -F demo build",
+    "sandbox": "pnpm -F sandbox dev",
     "docs:dev": "pnpm -F docs dev",
     "docs:build": "pnpm -F docs build",
     "ci:publish": "changeset publish",
